server: Cleanup how we bootstrap HTTP and socket server
Change-type: patch
