+++ lib/server/bootstrap.js
@@ -27,10 +27,8 @@ const actionServer = require('./action-server')
 const errorReporter = require('./error-reporter')
 const configuration = require('./configuration')
 const cardLoader = require('./card-loader')
-
-const {
-	createWebServer
-} = require('./web-server')
+const http = require('./http')
+const socket = require('./socket')
 
 module.exports = async (context) => {
 	const jellyfish = await core.create(context, {
@@ -204,13 +202,16 @@ module.exports = async (context) => {
 	await actionServer.flush(
 		mainWorkerContext, jellyfish, jellyfish.sessions.admin, workerManager, queue)
 
-	const server = await createWebServer(context, jellyfish, queue, configuration)
+	const webServer = await http(jellyfish, queue, configuration)
+	const socketServer = socket(jellyfish, webServer.server)
+	await webServer.start()
 	return {
 		jellyfish,
-		port: server.port,
 		queue,
+		port: webServer.port,
 		close: async () => {
-			await server.stop()
+			socketServer.close()
+			await webServer.stop()
 			await workerManager.stop(mainWorkerContext)
 			await jellyfish.disconnect(context)
 		}
